se agrega correcion de campo en varia tablas

+++ src/main/java/com/TicoTaco/TicoTaco/Repository/AdminRestauranteRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdminRestauranteRepository extends JpaRepository<AdminRestauranteModel, Integer> {
+public interface AdminRestauranteRepository extends JpaRepository<AdminRestauranteModel, Long> {
 }
